chore: adjust to separation between Certificate and CertificationRequest
