+++ packages/origin-backend-client-mocks/package.json
@@ -23,6 +23,7 @@
         "@energyweb/origin-backend-client": "6.2.1",
         "@energyweb/origin-backend-core": "4.2.0",
         "@energyweb/utils-general": "9.1.5",
+        "moment": "2.24.0",
         "moment-range": "4.0.2"
     },
     "files": [
