+++ packages/origin-backend-client-mocks/src/index.ts
@@ -1,7 +1,7 @@
-export * from './PreciseProofClientMock';
 export * from './ConfigurationClientMock';
 export * from './UserClientMock';
 export * from './OrganizationClientMock';
 export * from './DeviceClientMock';
 export * from './OffChainDataSourceMock';
 export * from './CertificateClientMock';
+export * from './CertificationRequestClientMock';
