+++ packages/origin-ui-core/src/components/OriginConfigurationContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, ReactNode } from 'react';
 import { createMuiTheme, Theme } from '@material-ui/core';
-import { OriginFeature } from '@energyweb/utils-general';
+import { OriginFeature, allOriginFeatures } from '@energyweb/utils-general';
 import i18n from 'i18next';
 import ICU from 'i18next-icu';
 import {
@@ -13,7 +13,6 @@ import variables from '../styles/variables.scss';
 import { OriginGenericLogo } from './icons/OriginGenericLogo';
 import { setTimeFormatLanguage } from '../utils/time';
 import { plPL, enUS } from '@material-ui/core/locale';
-import { getEnumValues } from '..';
 
 export interface IOriginStyleConfig {
     PRIMARY_COLOR: string;
@@ -245,7 +244,7 @@ export function createOriginConfiguration(configuration: Partial<IOriginConfigur
         materialTheme: createMaterialThemeForOrigin(DEFAULT_STYLE_CONFIG, storedLanguage),
         defaultLanguage: 'en',
         language: storedLanguage,
-        enabledFeatures: getEnumValues(OriginFeature)
+        enabledFeatures: allOriginFeatures
     };
 
     const newConfiguration: IOriginConfiguration = {
